res.send login functions
When there is an error at login, the error message is rendered and the login page disappears, causing the user to have to reload the website to try again. Instead, a simple function can be created and the 'alert' system can be used. The 'alert' function would cause a pop-up window to appear in browser, displaying a message to the user. This solves the problem of the website needing to be reloaded.